python-service/requirements.txt
flask>=3.1.2
flask-restful>=0.3.10
flask-cors>=6.0.1
requests>=2.32.5
python-dotenv>=1.1.1
pymongo>=4.14.1
pytest>=8.4.1